Move emoji-related utilities to new utils module
