+++ src/services/cxtie.service.ts
@@ -1,12 +1,15 @@
 import { Message } from "discord.js";
-import getLogger from "../logger";
-import { formatContext } from "../utils/logging.utils";
-import { parseCustomEmojis } from "../utils/markdown.utils";
 
-const SUP_EMOJI_ID = "1171056612761403413"
-const SLAY_EMOJI_ID = "1176908139896000623";
+import getLogger from "../logger";
+import { parseCustomEmojis } from "../utils/emojis.utils";
+import { formatContext } from "../utils/logging.utils";
 
 const log = getLogger(__filename);
+
+// NOTE: These emojis are from outside yung kai world so they are not part of
+// our custom emojis mapping.
+const SUP_EMOJI_ID = "1171056612761403413"
+const SLAY_EMOJI_ID = "1176908139896000623";
 
 export class CxtieService {
   public static INIT_TIMER_REACT_CHANCE = 0.05;
